backend/requirements.txt
fastapi~=0.104.0
SQLAlchemy~=2.0.22
passlib~=1.7.4
pydantic~=2.4.2
colorama~=0.4.6
PyYAML~=6.0.1
uvicorn~=0.23.2
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart